+++ src/index.ts
@@ -1,11 +1,22 @@
-export * from './assemble';
-export * from './chosungIncludes';
-export * from './combineHangulCharacter';
-export * from './convertQwertyToHangulAlphabet';
-export * from './disassemble';
-export * from './disassembleCompleteHangulCharacter';
-export * from './hangulIncludes';
-export * from './josa';
-export * from './removeLastHangulCharacter';
-export * from './typeHangul';
-export * from './utils';
+export { assembleHangul } from './assemble';
+export { chosungIncludes } from './chosungIncludes';
+export { combineHangulCharacter, combineVowels, curriedCombineHangulCharacter } from './combineHangulCharacter';
+export { convertQwertyToHangul, convertQwertyToHangulAlphabet } from './convertQwertyToHangulAlphabet';
+export { disassembleHangul, disassembleHangulToGroups } from './disassemble';
+export { disassembleCompleteHangulCharacter } from './disassembleCompleteHangulCharacter';
+export { hangulIncludes } from './hangulIncludes';
+export { josa } from './josa';
+export { removeLastHangulCharacter } from './removeLastHangulCharacter';
+export {
+  canBeChosung,
+  canBeJongsung,
+  canBeJungsung,
+  getChosung,
+  hasBatchim,
+  hasProperty,
+  hasSingleBatchim,
+  hasValueInReadOnlyStringList,
+} from './utils';
+export { extractHangul } from './extractHangul';
+export { acronymizeHangul } from './acronymizeHangul';
+export { typeHangul, getTypewriterHangul } from './typeHangul';